chore: Fractal renaming - Graph
