+++ .infra/rdev/values.yaml
@@ -2,7 +2,7 @@ stack:
   services:
     explorer:
       image:
-        tag: sha-965c3509
+        tag: sha-24127253
       replicaCount: 1
       env:
         # env vars common to all deployment stages
