+++ client/src/components/Graph/Graph.tsx
@@ -32,8 +32,8 @@ import { DatasetUnsMetadata } from "common/types/entities";
 import { getFeatureFlag } from "util/featureFlags/featureFlags";
 import { FEATURES } from "util/featureFlags/features";
 import { LassoFunctionWithAttributes } from "./setupLasso";
-import CentroidLabels from "./overlays/centroidLabels";
-import GraphOverlayLayer from "./overlays/graphOverlayLayer";
+import { CentroidLabels } from "./overlays/CentroidLabels/CentroidLabels";
+import { GraphOverlayLayer } from "./overlays/GraphOverlayLayer/GraphOverlayLayer";
 import * as globals from "~/globals";
 import _drawPoints from "./drawPointsRegl";
 import { setupBrush, setupLasso } from "./setupSVGandBrush";
