+++ client/src/components/Graph/constants.ts
