+++ client/src/components/Graph/drawPointsRegl.ts
